// client/src/types.ts
export interface LandingData {
  _id?: string;
  // Hero
  heroTitle: string;
  heroSub: string;
  heroImg: string;
  experienceYears: string;
  // About
  aboutBadge: string;
  aboutTitle: string;
  aboutText1: string;
  aboutText2: string;
  statYears: string;
  statProjects: string;
  aboutImg: string;
}